update: music機能を削除
update: whats_newにdpy統合を記載
fix: dpy_monkeyを読み込めるように
update: bot.add_cogのオーバーライドをモンキーパッチではなくした。
